requirements.txt
streamlit
google-generativeai
llama-index
pypdf2
reportlab
langchain
langgraph
gitpython
python-dotenv
asyncio